Refactor names, and remove warnings from vEB trees.
