+++ src/System.Collections.Algorithms/System.Collection.Algorithms.Tests/VanEmdeBoastTrees/VanEmdeBoasTree16Tests.cs
@@ -1,41 +1,40 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Xunit;
-
-namespace System.Collections.Algorithms.Tests.VebTree
+namespace System.Collections.Algorithms.Tests.VanEmdeBoasTrees
 {
-    public class VebTree32Tests
+    public class VanEmdeBoasTree16Tests
     {
         [Fact]
         public void GivenVebTreePopulatedWithOddNumbersWhenFindReturnsCorrectValue()
         {
-            var tree = new VebTree32();
-            for (uint i = 0; i < 1 << 8; i++)
+            var tree = new VanEmdeBoasTree16();
+            for (int i = 0; i <= ushort.MaxValue; i++)
                 if (i % 2 == 1)
-                    tree.Add(i);
-            for (uint i = 0; i < 1 << 8; i++)
-                Assert.Equal(i % 2 == 1, tree.Find(i));
+                    tree.Add((ushort)i);
+            for (int i = 0; i <= ushort.MaxValue; i++)
+                Assert.Equal(i % 2 == 1, tree.Find((ushort)i));
         }
 
         [Fact]
         public void GivenVebTreePopulatedWithEvenNumbersWhenFindReturnsCorrectValue()
         {
-            var tree = new VebTree32();
-            for (uint i = 0; i < 1 << 8; i++)
+            var tree = new VanEmdeBoasTree16();
+            for (int i = 0; i <= ushort.MaxValue; i++)
                 if (i % 2 == 0)
-                    tree.Add(i);
-            for (uint i = 0; i < 1 << 4; i++)
-                Assert.Equal(i % 2 == 0, tree.Find(i));
+                    tree.Add((ushort)i);
+            for (int i = 0; i <= ushort.MaxValue; i++)
+                Assert.Equal(i % 2 == 0, tree.Find((ushort)i));
         }
 
         [Fact]
         public void GivenEmptyVebTreeWhenFindThenAlwaysReturnFalse()
         {
-            var tree = new VebTree32();
+            var tree = new VanEmdeBoasTree16();
             var rand = new Random();
-            for (uint i = 0; i < 1 << 8; i++)
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
+                var elem = (ushort)rand.Next(1 << 8);
                 Assert.False(tree.Find(elem));
             }
         }
@@ -43,11 +42,11 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenTreeWhenPopulateItThenCountReturnsAmountOfAddedElements()
         {
-            var tree = new VebTree32();
+            var tree = new VanEmdeBoasTree16();
             Assert.Equal(0U, tree.Count);
-            for (uint i = 0; i < 1 << 17; i++)
+            for (uint i = 0; i <= ushort.MaxValue; i++)
             {
-                tree.Add(i);
+                tree.Add((ushort)i);
                 Assert.Equal(i + 1, tree.Count);
             }
         }
@@ -55,12 +54,12 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomCollectionWhenAddingElementsToTreeThenAddReturnsSameResultAsAddingToSet()
         {
-            var tree = new VebTree32();
-            var rand = new Random(1);
+            var tree = new VanEmdeBoasTree16();
+            var rand = new Random();
             var set = new HashSet<uint>();
-            for (int i = 0; i < 1 << 17; i++)
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
+                var elem = (ushort)rand.Next(1 << 16);
                 var add = set.Add(elem);
                 Assert.Equal(add, tree.Add(elem));
             }
@@ -68,12 +67,12 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenPopulatedVebTreeWhenFindThenAlwaysReturnTrue()
         {
-            var tree = new VebTree32();
+            var tree = new VanEmdeBoasTree16();
             var rand = new Random();
-            var set = new HashSet<uint>();
-            for (int i = 0; i < 1 << 20; i++)
+            var set = new HashSet<ushort>();
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
+                var elem = (ushort)rand.Next(1 << 16);
                 set.Add(elem);
                 tree.Add(elem);
             }
@@ -86,30 +85,30 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenVebTreeWhenAddSameNumberThenReturnFalse()
         {
-            var tree = new VebTree32();
-            for (uint i = 0; i < 1 << 17; i++)
+            var tree = new VanEmdeBoasTree16();
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                Assert.True(tree.Add(i));
-                Assert.False(tree.Add(i));
+                Assert.True(tree.Add((ushort)i));
+                Assert.False(tree.Add((ushort)i));
             }
 
-            tree = new VebTree32();
-            for (uint i = (1 << 18) - 1; i > 0; i--)
+            tree = new VanEmdeBoasTree16();
+            for (int i = ushort.MaxValue; i > 0; i--)
             {
-                Assert.True(tree.Add(i));
-                Assert.False(tree.Add(i));
+                Assert.True(tree.Add((ushort)i));
+                Assert.False(tree.Add((ushort)i));
             }
         }
 
         [Fact]
         public void GivenRandomPopulatedTreeWhenTryGetNextReturnsCorrectValue()
         {
-            var tree = new VebTree32();
+            var tree = new VanEmdeBoasTree16();
             var rand = new Random();
-            var list = new List<uint>();
-            for (int i = 0; i < 1 << 17; i++)
+            var list = new List<ushort>();
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
+                var elem = (ushort)rand.Next(1 << 16);
                 list.Add(elem);
                 tree.Add(elem);
             }
@@ -117,87 +116,87 @@ namespace System.Collections.Algorithms.Tests.VebTree
             var arr = list.OrderBy(x => x).Distinct().ToArray();
             if (arr[0] != 0)
             {
-                Assert.True(tree.TryGetNext(0, out uint res));
+                Assert.True(tree.TryGetNext(0, out ushort res));
                 Assert.Equal(arr[0], res);
             }
             for (int i = 1; i < arr.Length - 1; i++)
             {
                 if (arr[i] - arr[i - 1] > 1)
                 {
-                    Assert.True(tree.TryGetNext((arr[i] - arr[i - 1]) / 2 + arr[i - 1], out uint next));
+                    Assert.True(tree.TryGetNext((ushort)((arr[i] - arr[i - 1]) / 2 + arr[i - 1]), out ushort next));
                     Assert.Equal(arr[i], next);
                 }
-                Assert.True(tree.TryGetNext(arr[i - 1], out uint res));
+                Assert.True(tree.TryGetNext(arr[i - 1], out ushort res));
                 Assert.Equal(arr[i], res);
             }
-            Assert.False(tree.TryGetNext(arr[arr.Length - 1], out uint _));
+            Assert.False(tree.TryGetNext(arr[arr.Length - 1], out ushort _));
         }
 
         [Fact]
         public void GivenEmptyTreeWhenTryGetNextReturnsFalse()
         {
-            var tree = new VebTree32();
-            Assert.False(tree.TryGetNext(uint.MinValue, out uint _));
+            var tree = new VanEmdeBoasTree16();
+            Assert.False(tree.TryGetNext(ushort.MinValue, out ushort _));
         }
 
         [Fact]
         public void GivenEmptyTreeWhenTryGetPrevReturnsFalse()
         {
-            var tree = new VebTree32();
-            Assert.False(tree.TryGetPrevious(uint.MaxValue, out uint _));
+            var tree = new VanEmdeBoasTree16();
+            Assert.False(tree.TryGetPrevious(ushort.MaxValue, out ushort _));
         }
 
         [Fact]
         public void GivenRandomPopulatedTreeWhenGetPrevReturnsCorrectValue()
         {
-            var tree = new VebTree32();
+            var tree = new VanEmdeBoasTree16();
             var rand = new Random();
-            var list = new List<uint>();
-            for (int i = 0; i < 1 << 17; i++)
+            var list = new List<ushort>();
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
+                var elem = (ushort)rand.Next(1 << 16);
                 list.Add(elem);
                 tree.Add(elem);
             }
             var arr = list.OrderByDescending(x => x).Distinct().ToArray();
-            if (arr[0] != uint.MaxValue)
+            if (arr[0] != ushort.MaxValue)
             {
-                Assert.True(tree.TryGetPrevious(uint.MaxValue, out uint res));
+                Assert.True(tree.TryGetPrevious(ushort.MaxValue, out ushort res));
                 Assert.Equal(arr[0], res);
             }
             for (int i = 1; i < arr.Length - 1; i++)
             {
                 if (arr[i - 1] - arr[i] > 1)
                 {
-                    Assert.True(tree.TryGetPrevious((arr[i - 1] - arr[i]) / 2 + arr[i], out uint prev));
+                    Assert.True(tree.TryGetPrevious((ushort)((arr[i - 1] - arr[i]) / 2 + arr[i]), out ushort prev));
                     Assert.Equal(arr[i], prev);
                 }
-                Assert.True(tree.TryGetPrevious(arr[i - 1], out uint res));
+                Assert.True(tree.TryGetPrevious(arr[i - 1], out ushort res));
                 Assert.Equal(arr[i], res);
             }
-            Assert.False(tree.TryGetPrevious(arr[arr.Length - 1], out uint _));
+            Assert.False(tree.TryGetPrevious(arr[arr.Length - 1], out ushort _));
         }
 
         [Fact]
         public void GivenFullVebTreeWhenRemoveNumberThenFindCantFindIt()
         {
-            var tree = new VebTree32();
-            for (uint i = 0; i < 1 << 17; i++)
+            var tree = new VanEmdeBoasTree16();
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                tree.Add(i);
+                tree.Add((ushort)i);
             }
-            for (uint i = 0; i < 1 << 17; i++)
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                tree.Remove(i);
-                Assert.False(tree.Find(i));
+                tree.Remove((ushort)i);
+                Assert.False(tree.Find((ushort)i));
             }
-            for (uint i = 0; i < 1 << 17; i++)
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                tree.Add(i);
+                tree.Add((ushort)i);
             }
-            for (uint i = 0; i < 1 << 17; i++)
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                var toRemove = (1 << 17) - 1 - i;
+                var toRemove = (ushort)(ushort.MaxValue - i);
                 tree.Remove(toRemove);
                 Assert.False(tree.Find(toRemove));
             }
@@ -206,19 +205,19 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomFilledTreeWhenRemoveThenMatchesArrayState()
         {
-            var tree = new VebTree32();
+            var tree = new VanEmdeBoasTree16();
             var rand = new Random();
-            var arr = new bool[1 << 17];
-            for (int i = 0; i < 1 << 7; i++)
+            var arr = new bool[1 << 16];
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                var elem = (uint)rand.Next(1 << 17);
+                var elem = (ushort)rand.Next(1 << 16);
                 arr[elem] = true;
                 tree.Add(elem);
             }
 
-            for (int i = 0; i < 1 << 7; i++)
+            for (int i = 0; i <= ushort.MaxValue; i++)
             {
-                var elem = (uint)rand.Next(1 << 17);
+                var elem = (ushort)rand.Next(1 << 16);
                 Assert.Equal(arr[elem], tree.Remove(elem));
                 arr[elem] = false;
             }
@@ -227,13 +226,13 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomFilledTreeWhenCallMaxAndMinThenReturnProperValues()
         {
-            var tree = new VebTree32();
+            var tree = new VanEmdeBoasTree16();
             var rand = new Random();
-            uint min = uint.MaxValue;
-            uint max = uint.MinValue;
+            ushort min = ushort.MaxValue;
+            ushort max = ushort.MinValue;
             for (int i = 0; i < 1000; i++)
             {
-                var value = (uint)rand.Next();
+                var value = (ushort)rand.Next(1 << 16);
                 if (min > value) min = value;
                 if (max < value) max = value;
                 tree.Add(value);
