+++ src/System.Collections.Algorithms/System.Collection.Algorithms.Tests/VanEmdeBoastTrees/VanEmdeBoasTree32Tests.cs
@@ -2,40 +2,40 @@
 using System.Linq;
 using Xunit;
 
-namespace System.Collections.Algorithms.Tests.VebTree
+namespace System.Collections.Algorithms.Tests.VanEmdeBoasTrees
 {
-    public class VebTree64Tests
+    public class VanEmdeBoasTree32Tests
     {
         [Fact]
         public void GivenVebTreePopulatedWithOddNumbersWhenFindReturnsCorrectValue()
         {
-            var tree = new VebTree64();
-            for (ulong i = 0; i < 1 << 8; i++)
+            var tree = new VanEmdeBoasTree32();
+            for (uint i = 0; i < 1 << 8; i++)
                 if (i % 2 == 1)
                     tree.Add(i);
-            for (ulong i = 0; i < 1 << 8; i++)
+            for (uint i = 0; i < 1 << 8; i++)
                 Assert.Equal(i % 2 == 1, tree.Find(i));
         }
 
         [Fact]
         public void GivenVebTreePopulatedWithEvenNumbersWhenFindReturnsCorrectValue()
         {
-            var tree = new VebTree64();
-            for (ulong i = 0; i < 1 << 8; i++)
+            var tree = new VanEmdeBoasTree32();
+            for (uint i = 0; i < 1 << 8; i++)
                 if (i % 2 == 0)
                     tree.Add(i);
-            for (ulong i = 0; i < 1 << 4; i++)
+            for (uint i = 0; i < 1 << 4; i++)
                 Assert.Equal(i % 2 == 0, tree.Find(i));
         }
 
         [Fact]
         public void GivenEmptyVebTreeWhenFindThenAlwaysReturnFalse()
         {
-            var tree = new VebTree64();
+            var tree = new VanEmdeBoasTree32();
             var rand = new Random();
-            for (ulong i = 0; i < 1 << 8; i++)
+            for (uint i = 0; i < 1 << 8; i++)
             {
-                var elem = (ulong)(rand.Next(1 << 30)) << 2 | (ulong)(rand.Next(1 << 30));
+                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
                 Assert.False(tree.Find(elem));
             }
         }
@@ -43,11 +43,11 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenTreeWhenPopulateItThenCountReturnsAmountOfAddedElements()
         {
-            var tree = new VebTree64();
+            var tree = new VanEmdeBoasTree32();
             Assert.Equal(0U, tree.Count);
-            for (ulong i = 0; i < 1 << 17; i++)
+            for (uint i = 0; i < 1 << 17; i++)
             {
-                tree.Add(i << 20);
+                tree.Add(i);
                 Assert.Equal(i + 1, tree.Count);
             }
         }
@@ -55,12 +55,12 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomCollectionWhenAddingElementsToTreeThenAddReturnsSameResultAsAddingToSet()
         {
-            var tree = new VebTree64();
-            var rand = new Random();
-            var set = new HashSet<ulong>();
+            var tree = new VanEmdeBoasTree32();
+            var rand = new Random(1);
+            var set = new HashSet<uint>();
             for (int i = 0; i < 1 << 17; i++)
             {
-                var elem = (ulong)(rand.Next(1 << 30)) << 20 | (ulong)(rand.Next(1 << 30));
+                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
                 var add = set.Add(elem);
                 Assert.Equal(add, tree.Add(elem));
             }
@@ -68,12 +68,12 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenPopulatedVebTreeWhenFindThenAlwaysReturnTrue()
         {
-            var tree = new VebTree64();
+            var tree = new VanEmdeBoasTree32();
             var rand = new Random();
-            var set = new HashSet<ulong>();
-            for (int i = 0; i < 1 << 20; i++)
+            var set = new HashSet<uint>();
+            for (int i = 0; i < 1 << 10; i++)
             {
-                var elem = (ulong)(rand.Next(1 << 30)) << 20 | (ulong)(rand.Next(1 << 30));
+                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
                 set.Add(elem);
                 tree.Add(elem);
             }
@@ -86,30 +86,30 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenVebTreeWhenAddSameNumberThenReturnFalse()
         {
-            var tree = new VebTree64();
-            for (ulong i = 0; i < 1 << 17; i++)
+            var tree = new VanEmdeBoasTree32();
+            for (uint i = 0; i < 1 << 17; i++)
             {
-                Assert.True(tree.Add(i << 20));
-                Assert.False(tree.Add(i << 20));
+                Assert.True(tree.Add(i));
+                Assert.False(tree.Add(i));
             }
 
-            tree = new VebTree64();
-            for (ulong i = (1 << 18) - 1; i > 0; i--)
+            tree = new VanEmdeBoasTree32();
+            for (uint i = (1 << 18) - 1; i > 0; i--)
             {
-                Assert.True(tree.Add(i << 20));
-                Assert.False(tree.Add(i << 20));
+                Assert.True(tree.Add(i));
+                Assert.False(tree.Add(i));
             }
         }
 
         [Fact]
         public void GivenRandomPopulatedTreeWhenTryGetNextReturnsCorrectValue()
         {
-            var tree = new VebTree64();
+            var tree = new VanEmdeBoasTree32();
             var rand = new Random();
-            var list = new List<ulong>();
+            var list = new List<uint>();
             for (int i = 0; i < 1 << 17; i++)
             {
-                var elem = (ulong)(rand.Next(1 << 30)) << 2 | (ulong)(rand.Next(1 << 30));
+                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
                 list.Add(elem);
                 tree.Add(elem);
             }
@@ -117,87 +117,87 @@ namespace System.Collections.Algorithms.Tests.VebTree
             var arr = list.OrderBy(x => x).Distinct().ToArray();
             if (arr[0] != 0)
             {
-                Assert.True(tree.TryGetNext(0, out ulong res));
+                Assert.True(tree.TryGetNext(0, out uint res));
                 Assert.Equal(arr[0], res);
             }
             for (int i = 1; i < arr.Length - 1; i++)
             {
                 if (arr[i] - arr[i - 1] > 1)
                 {
-                    Assert.True(tree.TryGetNext((arr[i] - arr[i - 1]) / 2 + arr[i - 1], out ulong next));
+                    Assert.True(tree.TryGetNext((arr[i] - arr[i - 1]) / 2 + arr[i - 1], out uint next));
                     Assert.Equal(arr[i], next);
                 }
-                Assert.True(tree.TryGetNext(arr[i - 1], out ulong res));
+                Assert.True(tree.TryGetNext(arr[i - 1], out uint res));
                 Assert.Equal(arr[i], res);
             }
-            Assert.False(tree.TryGetNext(arr[arr.Length - 1], out ulong _));
+            Assert.False(tree.TryGetNext(arr[arr.Length - 1], out uint _));
         }
 
         [Fact]
         public void GivenEmptyTreeWhenTryGetNextReturnsFalse()
         {
-            var tree = new VebTree64();
-            Assert.False(tree.TryGetNext(ulong.MinValue, out ulong _));
+            var tree = new VanEmdeBoasTree32();
+            Assert.False(tree.TryGetNext(uint.MinValue, out uint _));
         }
 
         [Fact]
         public void GivenEmptyTreeWhenTryGetPrevReturnsFalse()
         {
-            var tree = new VebTree64();
-            Assert.False(tree.TryGetPrevious(ulong.MaxValue, out ulong _));
+            var tree = new VanEmdeBoasTree32();
+            Assert.False(tree.TryGetPrevious(uint.MaxValue, out uint _));
         }
 
         [Fact]
         public void GivenRandomPopulatedTreeWhenGetPrevReturnsCorrectValue()
         {
-            var tree = new VebTree64();
+            var tree = new VanEmdeBoasTree32();
             var rand = new Random();
-            var list = new List<ulong>();
+            var list = new List<uint>();
             for (int i = 0; i < 1 << 17; i++)
             {
-                var elem = (ulong)(rand.Next(1 << 30)) << 2 | (ulong)(rand.Next(1 << 30));
+                var elem = (uint)(rand.Next(1 << 30)) << 2 | (uint)(rand.Next(1 << 2));
                 list.Add(elem);
                 tree.Add(elem);
             }
             var arr = list.OrderByDescending(x => x).Distinct().ToArray();
-            if (arr[0] != ulong.MaxValue)
+            if (arr[0] != uint.MaxValue)
             {
-                Assert.True(tree.TryGetPrevious(ulong.MaxValue, out ulong res));
+                Assert.True(tree.TryGetPrevious(uint.MaxValue, out uint res));
                 Assert.Equal(arr[0], res);
             }
             for (int i = 1; i < arr.Length - 1; i++)
             {
                 if (arr[i - 1] - arr[i] > 1)
                 {
-                    Assert.True(tree.TryGetPrevious((arr[i - 1] - arr[i]) / 2 + arr[i], out ulong prev));
+                    Assert.True(tree.TryGetPrevious((arr[i - 1] - arr[i]) / 2 + arr[i], out uint prev));
                     Assert.Equal(arr[i], prev);
                 }
-                Assert.True(tree.TryGetPrevious(arr[i - 1], out ulong res));
+                Assert.True(tree.TryGetPrevious(arr[i - 1], out uint res));
                 Assert.Equal(arr[i], res);
             }
-            Assert.False(tree.TryGetPrevious(arr[arr.Length - 1], out ulong _));
+            Assert.False(tree.TryGetPrevious(arr[arr.Length - 1], out uint _));
         }
 
         [Fact]
         public void GivenFullVebTreeWhenRemoveNumberThenFindCantFindIt()
         {
-            var tree = new VebTree64();
-            for (ulong i = 0; i < 1 << 17; i++)
+            var tree = new VanEmdeBoasTree32();
+            for (uint i = 0; i < 1 << 17; i++)
             {
-                tree.Add(i << 20);
+                tree.Add(i);
             }
-            for (ulong i = 0; i < 1 << 17; i++)
+            for (uint i = 0; i < 1 << 17; i++)
             {
-                tree.Remove(i << 20);
+                tree.Remove(i);
                 Assert.False(tree.Find(i));
             }
-            for (ulong i = 0; i < 1 << 17; i++)
+            for (uint i = 0; i < 1 << 17; i++)
             {
-                tree.Add(i << 20);
+                tree.Add(i);
             }
-            for (ulong i = 0; i < 1 << 17; i++)
+            for (uint i = 0; i < 1 << 17; i++)
             {
-                var toRemove = (1 << 37) - 1 - (i << 20);
+                var toRemove = (1 << 17) - 1 - i;
                 tree.Remove(toRemove);
                 Assert.False(tree.Find(toRemove));
             }
@@ -206,19 +206,19 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomFilledTreeWhenRemoveThenMatchesArrayState()
         {
-            var tree = new VebTree64();
+            var tree = new VanEmdeBoasTree32();
             var rand = new Random();
             var arr = new bool[1 << 17];
             for (int i = 0; i < 1 << 7; i++)
             {
-                var elem = (ulong)rand.Next(1 << 17);
+                var elem = (uint)rand.Next(1 << 17);
                 arr[elem] = true;
                 tree.Add(elem);
             }
 
             for (int i = 0; i < 1 << 7; i++)
             {
-                var elem = (ulong)rand.Next(1 << 17);
+                var elem = (uint)rand.Next(1 << 17);
                 Assert.Equal(arr[elem], tree.Remove(elem));
                 arr[elem] = false;
             }
@@ -227,13 +227,13 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomFilledTreeWhenCallMaxAndMinThenReturnProperValues()
         {
-            var tree = new VebTree64();
+            var tree = new VanEmdeBoasTree32();
             var rand = new Random();
-            ulong min = ulong.MaxValue;
-            ulong max = ulong.MinValue;
+            uint min = uint.MaxValue;
+            uint max = uint.MinValue;
             for (int i = 0; i < 1000; i++)
             {
-                var value = (ulong)rand.Next();
+                var value = (uint)rand.Next();
                 if (min > value) min = value;
                 if (max < value) max = value;
                 tree.Add(value);
