+++ src/System.Collections.Algorithms/System.Collection.Algorithms.Tests/VanEmdeBoastTrees/VanEmdeBoasTree64Tests.cs
@@ -1,42 +1,41 @@
-﻿using System;
-using System.Collections.Algorithms;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using Xunit;
-namespace System.Collections.Algorithms.Tests.VebTree
+
+namespace System.Collections.Algorithms.Tests.VanEmdeBoasTrees
 {
-    public class VebTree16Tests
+    public class VanEmdeBoasTree64Tests
     {
         [Fact]
         public void GivenVebTreePopulatedWithOddNumbersWhenFindReturnsCorrectValue()
         {
-            var tree = new VebTree16();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var tree = new VanEmdeBoasTree64();
+            for (ulong i = 0; i < 1 << 8; i++)
                 if (i % 2 == 1)
-                    tree.Add((ushort)i);
-            for (int i = 0; i <= ushort.MaxValue; i++)
-                Assert.Equal(i % 2 == 1, tree.Find((ushort)i));
+                    tree.Add(i);
+            for (ulong i = 0; i < 1 << 8; i++)
+                Assert.Equal(i % 2 == 1, tree.Find(i));
         }
 
         [Fact]
         public void GivenVebTreePopulatedWithEvenNumbersWhenFindReturnsCorrectValue()
         {
-            var tree = new VebTree16();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var tree = new VanEmdeBoasTree64();
+            for (ulong i = 0; i < 1 << 8; i++)
                 if (i % 2 == 0)
-                    tree.Add((ushort)i);
-            for (int i = 0; i <= ushort.MaxValue; i++)
-                Assert.Equal(i % 2 == 0, tree.Find((ushort)i));
+                    tree.Add(i);
+            for (ulong i = 0; i < 1 << 4; i++)
+                Assert.Equal(i % 2 == 0, tree.Find(i));
         }
 
         [Fact]
         public void GivenEmptyVebTreeWhenFindThenAlwaysReturnFalse()
         {
-            var tree = new VebTree16();
+            var tree = new VanEmdeBoasTree64();
             var rand = new Random();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            for (ulong i = 0; i < 1 << 8; i++)
             {
-                var elem = (ushort)rand.Next(1 << 8);
+                var elem = GetRandomUlong(rand);
                 Assert.False(tree.Find(elem));
             }
         }
@@ -44,11 +43,11 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenTreeWhenPopulateItThenCountReturnsAmountOfAddedElements()
         {
-            var tree = new VebTree16();
+            var tree = new VanEmdeBoasTree64();
             Assert.Equal(0U, tree.Count);
-            for (uint i = 0; i <= ushort.MaxValue; i++)
+            for (ulong i = 0; i < 1 << 17; i++)
             {
-                tree.Add((ushort)i);
+                tree.Add(i << 20);
                 Assert.Equal(i + 1, tree.Count);
             }
         }
@@ -56,12 +55,12 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomCollectionWhenAddingElementsToTreeThenAddReturnsSameResultAsAddingToSet()
         {
-            var tree = new VebTree16();
+            var tree = new VanEmdeBoasTree64();
             var rand = new Random();
-            var set = new HashSet<uint>();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var set = new HashSet<ulong>();
+            for (int i = 0; i < 1 << 17; i++)
             {
-                var elem = (ushort)rand.Next(1 << 16);
+                var elem = GetRandomUlong(rand);
                 var add = set.Add(elem);
                 Assert.Equal(add, tree.Add(elem));
             }
@@ -69,12 +68,12 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenPopulatedVebTreeWhenFindThenAlwaysReturnTrue()
         {
-            var tree = new VebTree16();
+            var tree = new VanEmdeBoasTree64();
             var rand = new Random();
-            var set = new HashSet<ushort>();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var set = new HashSet<ulong>();
+            for (int i = 0; i < 1 << 10; i++)
             {
-                var elem = (ushort)rand.Next(1 << 16);
+                var elem = GetRandomUlong(rand);
                 set.Add(elem);
                 tree.Add(elem);
             }
@@ -87,30 +86,30 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenVebTreeWhenAddSameNumberThenReturnFalse()
         {
-            var tree = new VebTree16();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var tree = new VanEmdeBoasTree64();
+            for (ulong i = 0; i < 1 << 10; i++)
             {
-                Assert.True(tree.Add((ushort)i));
-                Assert.False(tree.Add((ushort)i));
+                Assert.True(tree.Add(i << 20));
+                Assert.False(tree.Add(i << 20));
             }
 
-            tree = new VebTree16();
-            for (int i = ushort.MaxValue; i > 0; i--)
+            tree = new VanEmdeBoasTree64();
+            for (ulong i = (1 << 18) - 1; i > 0; i--)
             {
-                Assert.True(tree.Add((ushort)i));
-                Assert.False(tree.Add((ushort)i));
+                Assert.True(tree.Add(i << 20));
+                Assert.False(tree.Add(i << 20));
             }
         }
 
         [Fact]
         public void GivenRandomPopulatedTreeWhenTryGetNextReturnsCorrectValue()
         {
-            var tree = new VebTree16();
+            var tree = new VanEmdeBoasTree64();
             var rand = new Random();
-            var list = new List<ushort>();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var list = new List<ulong>();
+            for (int i = 0; i < 1 << 17; i++)
             {
-                var elem = (ushort)rand.Next(1 << 16);
+                var elem = GetRandomUlong(rand);
                 list.Add(elem);
                 tree.Add(elem);
             }
@@ -118,87 +117,87 @@ namespace System.Collections.Algorithms.Tests.VebTree
             var arr = list.OrderBy(x => x).Distinct().ToArray();
             if (arr[0] != 0)
             {
-                Assert.True(tree.TryGetNext(0, out ushort res));
+                Assert.True(tree.TryGetNext(0, out ulong res));
                 Assert.Equal(arr[0], res);
             }
             for (int i = 1; i < arr.Length - 1; i++)
             {
                 if (arr[i] - arr[i - 1] > 1)
                 {
-                    Assert.True(tree.TryGetNext((ushort)((arr[i] - arr[i - 1]) / 2 + arr[i - 1]), out ushort next));
+                    Assert.True(tree.TryGetNext((arr[i] - arr[i - 1]) / 2 + arr[i - 1], out ulong next));
                     Assert.Equal(arr[i], next);
                 }
-                Assert.True(tree.TryGetNext(arr[i - 1], out ushort res));
+                Assert.True(tree.TryGetNext(arr[i - 1], out ulong res));
                 Assert.Equal(arr[i], res);
             }
-            Assert.False(tree.TryGetNext(arr[arr.Length - 1], out ushort _));
+            Assert.False(tree.TryGetNext(arr[arr.Length - 1], out ulong _));
         }
 
         [Fact]
         public void GivenEmptyTreeWhenTryGetNextReturnsFalse()
         {
-            var tree = new VebTree16();
-            Assert.False(tree.TryGetNext(ushort.MinValue, out ushort _));
+            var tree = new VanEmdeBoasTree64();
+            Assert.False(tree.TryGetNext(ulong.MinValue, out ulong _));
         }
 
         [Fact]
         public void GivenEmptyTreeWhenTryGetPrevReturnsFalse()
         {
-            var tree = new VebTree16();
-            Assert.False(tree.TryGetPrevious(ushort.MaxValue, out ushort _));
+            var tree = new VanEmdeBoasTree64();
+            Assert.False(tree.TryGetPrevious(ulong.MaxValue, out ulong _));
         }
 
         [Fact]
         public void GivenRandomPopulatedTreeWhenGetPrevReturnsCorrectValue()
         {
-            var tree = new VebTree16();
+            var tree = new VanEmdeBoasTree64();
             var rand = new Random();
-            var list = new List<ushort>();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var list = new List<ulong>();
+            for (int i = 0; i < 1 << 17; i++)
             {
-                var elem = (ushort)rand.Next(1 << 16);
+                var elem = GetRandomUlong(rand);
                 list.Add(elem);
                 tree.Add(elem);
             }
             var arr = list.OrderByDescending(x => x).Distinct().ToArray();
-            if (arr[0] != ushort.MaxValue)
+            if (arr[0] != ulong.MaxValue)
             {
-                Assert.True(tree.TryGetPrevious(ushort.MaxValue, out ushort res));
+                Assert.True(tree.TryGetPrevious(ulong.MaxValue, out ulong res));
                 Assert.Equal(arr[0], res);
             }
             for (int i = 1; i < arr.Length - 1; i++)
             {
                 if (arr[i - 1] - arr[i] > 1)
                 {
-                    Assert.True(tree.TryGetPrevious((ushort)((arr[i - 1] - arr[i]) / 2 + arr[i]), out ushort prev));
+                    Assert.True(tree.TryGetPrevious((arr[i - 1] - arr[i]) / 2 + arr[i], out ulong prev));
                     Assert.Equal(arr[i], prev);
                 }
-                Assert.True(tree.TryGetPrevious(arr[i - 1], out ushort res));
+                Assert.True(tree.TryGetPrevious(arr[i - 1], out ulong res));
                 Assert.Equal(arr[i], res);
             }
-            Assert.False(tree.TryGetPrevious(arr[arr.Length - 1], out ushort _));
+            Assert.False(tree.TryGetPrevious(arr[arr.Length - 1], out ulong _));
         }
 
         [Fact]
         public void GivenFullVebTreeWhenRemoveNumberThenFindCantFindIt()
         {
-            var tree = new VebTree16();
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var tree = new VanEmdeBoasTree64();
+            for (ulong i = 0; i < 1 << 17; i++)
             {
-                tree.Add((ushort)i);
+                tree.Add(i << 20);
             }
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            for (ulong i = 0; i < 1 << 17; i++)
             {
-                tree.Remove((ushort)i);
-                Assert.False(tree.Find((ushort)i));
+                tree.Remove(i << 20);
+                Assert.False(tree.Find(i));
             }
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            for (ulong i = 0; i < 1 << 17; i++)
             {
-                tree.Add((ushort)i);
+                tree.Add(i << 20);
             }
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            for (ulong i = 0; i < 1 << 17; i++)
             {
-                var toRemove = (ushort)(ushort.MaxValue - i);
+                var toRemove = (1 << 37) - 1 - (i << 20);
                 tree.Remove(toRemove);
                 Assert.False(tree.Find(toRemove));
             }
@@ -207,19 +206,19 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomFilledTreeWhenRemoveThenMatchesArrayState()
         {
-            var tree = new VebTree16();
+            var tree = new VanEmdeBoasTree64();
             var rand = new Random();
-            var arr = new bool[1 << 16];
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            var arr = new bool[1 << 17];
+            for (int i = 0; i < 1 << 7; i++)
             {
-                var elem = (ushort)rand.Next(1 << 16);
+                var elem = (ulong)rand.Next(1 << 17);
                 arr[elem] = true;
                 tree.Add(elem);
             }
 
-            for (int i = 0; i <= ushort.MaxValue; i++)
+            for (int i = 0; i < 1 << 7; i++)
             {
-                var elem = (ushort)rand.Next(1 << 16);
+                var elem = (ulong)rand.Next(1 << 17);
                 Assert.Equal(arr[elem], tree.Remove(elem));
                 arr[elem] = false;
             }
@@ -228,13 +227,13 @@ namespace System.Collections.Algorithms.Tests.VebTree
         [Fact]
         public void GivenRandomFilledTreeWhenCallMaxAndMinThenReturnProperValues()
         {
-            var tree = new VebTree16();
+            var tree = new VanEmdeBoasTree64();
             var rand = new Random();
-            ushort min = ushort.MaxValue;
-            ushort max = ushort.MinValue;
+            ulong min = ulong.MaxValue;
+            ulong max = ulong.MinValue;
             for (int i = 0; i < 1000; i++)
             {
-                var value = (ushort)rand.Next(1 << 16);
+                var value = GetRandomUlong(rand);
                 if (min > value) min = value;
                 if (max < value) max = value;
                 tree.Add(value);
@@ -243,5 +242,11 @@ namespace System.Collections.Algorithms.Tests.VebTree
             Assert.Equal(max, tree.Max);
         }
 
+        private ulong GetRandomUlong(Random rand)
+        {
+            var upper = ((ulong)rand.Next(1 << 30) << 31);
+            var lower = (ulong)rand.Next(1 << 30);
+            return upper | lower;
+        }
     }
 }
